adjustments in wax modeling; bug correction; new gui

+++ src/GradientCorrelations.h
@@ -29,10 +29,10 @@ const double g_in=32.174;            // Aceleração da gravidade em ft/s² (sis
 // Cálculo do fator de atrito usando diagrama de Moody
 void frictionFactor(double reynolds, double roughness, double& frictionFactorValue);
 
-long double interpolate(const std::vector<double>& xValues, const std::vector<double>& yValues,
+double interpolate(const std::vector<double>& xValues, const std::vector<double>& yValues,
                    double xTarget, int degree, int numPoints);
 
-long double interpolate2D(const std::vector<double>& verticalPoints,
+double interpolate2D(const std::vector<double>& verticalPoints,
                      const std::vector<double>& horizontalPoints,
                      const std::vector<double>& matrixValues,
                      int numVertical, int numHorizontal,
